wtt/requirements.txt
django==5.1.4
asgiref==3.8.1
sqlparse==0.5.3
psycopg2-binary==2.9.10
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.9.0
uvicorn==0.34.0
djangorestframework-simplejwt==5.3.1
pyjwt==2.10.1